AISLE Phase 6: frontend (all 8 screens) + backend read endpoints
Adds the Next.js 14 App Router frontend: Overview, Workflow, Themes (+
detail), Insights (+ detail with IQS radar chart and counter-evidence
panel), Quality (classifier report card + the real negative-control
verdict), Ask (question-pack shell, honest 404 fallback pending Phase 7),
Upload (drag-and-drop preview/commit against the Phase 2 backend), and
Admin (sources table, on-demand pipeline-stage triggers, a minimal
labelling widget for the Phase 3 golden set).

components/PercentCI.tsx is the enforcement mechanism for the brief's "no
naked percentage, anywhere" rule - every rate in the UI renders through it,
so n/denominator/CI can't be dropped without deliberately bypassing the
component.

Backend: new read-only routers (overview, runs, themes, insights, sources)
plus /admin/run/* synchronous pipeline triggers and /quality/negative-control
to surface the Phase 5 experiment result in the UI. Fixed a real bug found
while testing against the live stack: /insights and /quality/negative-control
threw psycopg.errors.IndeterminateDatatype on the frontend's actual
no-filter request (a bare "%s IS NULL OR col = %s" needs an explicit
::type cast when every call site passes None) - fixed and covered by a new
regression test file. Also fixed pgvector's raw ndarray leaking into the
theme-detail JSON response (unserialisable, and not needed by the UI).

Verified against the live stack, not just `next build`: ran the FastAPI
backend and Next.js dev server together, hit all 8 screens plus theme/
insight detail pages with Playwright, and confirmed real corpus data
renders (real Wilson CIs, the real IQS breakdown, the real negative-control
PASS verdict).

Honesty caveats in README: no shadcn/ui (plain Tailwind components
instead), /ask is a UI shell pending Phase 7, /admin run-triggers are
synchronous rather than queued.

85/85 backend tests passing (77 previous + 8 new API-router regression
tests).

+++ backend/aisle/eval/negative_control.py
@@ -37,7 +37,7 @@ def run_negative_control_check(run_id: int | None = None) -> dict:
             JOIN themes t ON t.id = ANY(i.theme_ids)
             JOIN theme_documents td ON td.theme_id = t.id
             JOIN documents d ON d.id = td.document_id
-            WHERE (%s IS NULL OR i.run_id = %s)
+            WHERE (%s::bigint IS NULL OR i.run_id = %s::bigint)
             GROUP BY i.id, i.title, i.grade, i.iqs_total, i.theme_ids
             """,
             (run_id, run_id),
